// frontend/src/lib/api/models/StudentAvailabilityRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StudentAvailabilityRead = {
    moduleId: string;
    studyDays: Array<string>;
    preferredWindow: string;
    maxStudyMinutesPerDay: number;
    availabilityVersion: number;
    updatedAt?: (string | null);
};